test(parser): move over the validator tests and some misc tests to parser
