+++ crates/aiken-lang/src/parser.rs
@@ -38,3 +38,24 @@ pub fn module(
 
     Ok((module, extra))
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::assert_module;
+
+    #[test]
+    fn windows_newline() {
+        assert_module!("use aiken/list\r\n");
+    }
+
+    #[test]
+    fn can_handle_comments_at_end_of_file() {
+        assert_module!(
+            r#"
+            use aiken
+
+            // some comment
+            // more comments"#
+        );
+    }
+}
