+++ crates/aiken-lang/src/parser/expr/if_else.rs
@@ -49,8 +49,6 @@ pub fn parser<'a>(
 
 #[cfg(test)]
 mod tests {
-    use chumsky::Parser;
-
     use crate::assert_expr;
 
     #[test]
